+++ esm/app.js
@@ -14,7 +14,7 @@ export class App {
         return this._events[name];
     }
     run(name, ...args) {
-        const subscribers = this._events[name] || [];
+        const subscribers = this.getSubscribers(name, this._events);
         console.assert(subscribers && subscribers.length > 0, 'No subscriber for event: ' + name);
         // Update the list of subscribers by pulling out those which will run once.
         // We must do this update prior to running any of the events in case they
@@ -28,7 +28,7 @@ export class App {
                 this.delay(name, fn, args, options);
             }
             else {
-                fn.apply(this, args);
+                Object.keys(options).length > 0 ? fn.apply(this, [...args, options]) : fn.apply(this, args);
             }
             return !sub.options.once;
         });
@@ -42,8 +42,28 @@ export class App {
             clearTimeout(options._t);
         options._t = setTimeout(() => {
             clearTimeout(options._t);
-            fn.apply(this, args);
+            Object.keys(options).length > 0 ? fn.apply(this, [...args, options]) : fn.apply(this, args);
         }, options.delay);
+    }
+    query(name, ...args) {
+        const subscribers = this.getSubscribers(name, this._events);
+        console.assert(subscribers && subscribers.length > 0, 'No subscriber for event: ' + name);
+        const promises = subscribers.map(sub => {
+            const { fn, options } = sub;
+            return Object.keys(options).length > 0 ? fn.apply(this, [...args, options]) : fn.apply(this, args);
+        });
+        return Promise.all(promises);
+    }
+    getSubscribers(name, events) {
+        const subscribers = this._events[name] || [];
+        const p = subscribers.length;
+        Object.keys(events).filter(evt => evt.endsWith('*') && name.startsWith(evt.replace('*', '')))
+            .sort((a, b) => b.length - a.length)
+            .forEach(evt => subscribers.push(...events[evt].map(sub => {
+            sub.options = Object.assign(Object.assign({}, sub.options), { event: name });
+            return sub;
+        })));
+        return subscribers;
     }
 }
 const AppRunVersions = 'AppRun-2';